backend/requirements.txt
asgiref==3.11.1
Django==6.0.3
django-cors-headers==4.9.0
djangorestframework==3.17.1
dotenv==0.9.9
gunicorn==25.2.0
packaging==26.0
psycopg2-binary==2.9.11
python-dotenv==1.2.2
sqlparse==0.5.5
